+++ mgt2e/module/documents/actor.mjs
@@ -1561,8 +1561,11 @@ export class MgT2Actor extends Actor {
         return this.effects.find(e => e.name === name);
     }
 
-    async setEffect(status, value, overlay, locked, css) {
+    async setEffect(status, value, overlay, locked, css, changes) {
         const statusEffect = CONFIG.statusEffects.find(e => e.id === status);
+        if (!changes) {
+            changes = [];
+        }
 
         if (!statusEffect) {
             ui.notifications.error(
@@ -1575,6 +1578,33 @@ export class MgT2Actor extends Actor {
         const effect = this.effects.find(e => e.name === name);
 
         if (value && effect) {
+            if (isNaN(parseInt(value))) {
+                return false;
+            } else if (effect.flags?.mgt2e?.value) {
+                // Passed in a numerical value. Need to +/- it.
+                let current = parseInt(effect.flags?.mgt2e?.value);
+                if (!isNaN(current) && current !== 0) {
+                    if (CONFIG.MGT2.STATUS_EFFECTS[status]?.replace) {
+                        current = parseInt(value);
+                    } else {
+                        current += parseInt(value);
+                    }
+                    effect.setFlag("mgt2e", "value", current);
+                    if (!CONFIG.MGT2.STATUS_EFFECTS[status]?.mono) {
+                        if (current < 0) {
+                            effect.setFlag("mgt2e", "css", "statusWarn");
+                        } else if (current > 0) {
+                            effect.setFlag("mgt2e", "css", "statusGood");
+                        }
+                    }
+                    if (effect.changes && effect.changes.length > 0) {
+                        effect.changes[0].value = current;
+                        effect.update({"changes": effect.changes});
+                    }
+                } if (current === 0) {
+                    effect.delete();
+                }
+            }
             return false;
         } else if (!value && !effect) {
             return false;
@@ -1582,7 +1612,7 @@ export class MgT2Actor extends Actor {
             await this.createEmbeddedDocuments("ActiveEffect", [{
                 name: name,
                 icon: statusEffect.img,
-                changes: [],
+                changes: changes,
                 statuses: [ status ],
                 flags: {
                     "core": {
@@ -1608,6 +1638,65 @@ export class MgT2Actor extends Actor {
         return true;
     }
 
+    addStatusEffect(status, value) {
+        switch (status) {
+            case "dead":
+                this.setDeadEffect(true);
+                break;
+            case "unconscious":
+                this.setUnconsciousEffect(true);
+                break;
+            case "injured":
+                this.setInjuredEffect(true);
+                break;
+            case "fear":
+                this.setFearEffect(true);
+                break;
+            case "stunned":
+                this.setStunnedEffect(value);
+                break;
+            case "needsFirstAid":
+                this.setFirstAidEffect(true);
+                true;
+            case "needsSurgery":
+                this.setSurgeryEffect(true)
+                break;
+            case "destroyed":
+                this.setDestroyedEffect(true);
+                break;
+            case "aware":
+                this.setAwareEffect(true);
+                break;
+            case "surprised":
+                this.setSurprisedEffect(true);
+                break;
+            case "fatigued":
+                this.setFatiguedEffect(true);
+                break;
+            case "inCover":
+                this.setInCoverEffect(value);
+                break;
+            case "prone":
+                this.setProneEffect(true);
+                break;
+            case "armour":
+                this.setArmourEffect(1);
+                break;
+            case "physical":
+                this.setPhysicalEffect(value);
+                break;
+            case "melee":
+                this.setMeleeEffect(value);
+                break;
+            case "gunCombat":
+                this.setGunCombatEffect(value);
+                break;
+            case "tactics":
+                this.setTacticsEffect(1);
+                break;
+        }
+    }
+
     setDeadEffect(value) {
         this.setEffect("dead", value, true, false,"Bad");
     }
@@ -1620,12 +1709,12 @@ export class MgT2Actor extends Actor {
         this.setEffect("injured", value,  false, false, "Warn");
     }
 
-    setFrightenedEffect(value) {
+    setFearEffect(value) {
         this.setEffect("fear", value,  false, false, "Warn");
     }
 
     setStunnedEffect(value) {
-        return this.setEffect("stun", value,  false, false, "Bad");
+        return this.setEffect("stunned", value,  false, false, "Bad");
     }
 
     setFirstAidEffect(value) {
@@ -1646,6 +1735,56 @@ export class MgT2Actor extends Actor {
 
     setVaccSuitEffect(value) {
         this.setEffect("vaccSuit", value,  false, true, "Warn");
+    }
+
+    setAwareEffect(value) {
+        this.setEffect("aware", value,  false, false, "Good");
+    }
+
+    setSurprisedEffect(value) {
+        this.setEffect("surprised", value,  false, false, "Warn");
+    }
+
+    setReactionEffect(value) {
+        this.setEffect("reaction", value,  false, false, "Warn");
+    }
+
+    setFatiguedEffect(value) {
+      this.setEffect("fatigued", value, false, false, "Warn",
+          [
+              { key: "system.modifiers.physical.effect", mode: 2, priority: 0, value: -2 }
+          ]);
+    }
+    setPhysicalEffect(value) {
+        let css= "Warn";
+        if (parseInt(value) > 0) {
+            css = "Good";
+        }
+        this.setEffect("physical", value, false, false, css,
+            [
+                { key: "system.modifiers.physical.effect", mode: 2, priority: 0, value: parseInt(value) }
+            ]);
+    }
+    setArmourEffect(value) {
+        this.setEffect("armour", value, false, false, "Good",
+            [
+                { key: "system.modifiers.armour.effect", mode: 2, priority: 0, value: parseInt(value) }
+            ]);
+    }
+    setInCoverEffect(value) {
+        this.setEffect("inCover", value, false, false, "Good",
+            [
+                { key: "system.modifiers.armour.effect", mode: 2, priority: 0, value: parseInt(value) }
+            ]);
+    }
+    setProneEffect(value) {
+        this.setEffect("prone", value, false, false, "Good");
+    }
+    setTacticsEffect(value) {
+        this.setEffect("tactics", value, false, false, "Good",
+            [
+                { key: "system.modifiers.initiative.effect", mode: 2, priority: 0, value: parseInt(value) }
+            ]);
     }
 
 }